Cleanup: Moving files from src/Microsoft.AspNet.Mvc.Razor/Razor/PreCompileViews -> src/Microsoft.AspNet.Mvc.Razor/Precompilation
