+++ src/Microsoft.AspNet.Mvc.Razor/Compilation/RazorFileHash.cs
@@ -5,9 +5,10 @@ using System;
 using System.Globalization;
 using System.IO;
 using Microsoft.AspNet.FileProviders;
+using Microsoft.AspNet.Mvc.Razor.Internal;
 using Microsoft.Framework.Internal;
 
-namespace Microsoft.AspNet.Mvc.Razor
+namespace Microsoft.AspNet.Mvc.Razor.Compilation
 {
     public static class RazorFileHash
     {
