+++ src/Microsoft.AspNet.Mvc.Razor/Precompilation/GeneratorResultExtensions.cs
@@ -8,7 +8,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.Framework.Internal;
 
-namespace Microsoft.AspNet.Mvc.Razor
+namespace Microsoft.AspNet.Mvc.Razor.Precompilation
 {
     public static class GeneratorResultExtensions
     {
