+++ src/Microsoft.AspNet.Mvc.Razor/Precompilation/PrecompilationCacheEntry.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using Microsoft.CodeAnalysis;
 using Microsoft.Framework.Internal;
 
-namespace Microsoft.AspNet.Mvc.Razor
+namespace Microsoft.AspNet.Mvc.Razor.Precompilation
 {
     /// <summary>
     /// An entry in the cache used by <see cref="RazorPreCompiler"/>.
