+++ src/Microsoft.AspNet.Mvc.Razor/Precompilation/RazorFileInfo.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.AspNet.Mvc.Razor
+namespace Microsoft.AspNet.Mvc.Razor.Precompilation
 {
     public class RazorFileInfo
     {
