+++ src/Microsoft.AspNet.Mvc.Razor/Precompilation/RazorFileInfoCollectionGenerator.cs
@@ -3,12 +3,13 @@
 
 using System.Globalization;
 using System.Text;
+using Microsoft.AspNet.Mvc.Razor.Compilation;
 using Microsoft.CodeAnalysis;
 using Microsoft.Framework.Internal;
 using Microsoft.Framework.Runtime;
 using Microsoft.Framework.Runtime.Roslyn;
 
-namespace Microsoft.AspNet.Mvc.Razor
+namespace Microsoft.AspNet.Mvc.Razor.Precompilation
 {
     public class RazorFileInfoCollectionGenerator
     {
