+++ src/Microsoft.AspNet.Mvc.Razor/VirtualPathRazorPageFactory.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Concurrent;
+using Microsoft.AspNet.Mvc.Razor.Compilation;
 using Microsoft.Framework.DependencyInjection;
 using Microsoft.Framework.Internal;
