+++ test/Microsoft.AspNet.Mvc.Razor.Test/Compilation/CompilationResultTest.cs
@@ -5,7 +5,7 @@ using Microsoft.Framework.Runtime;
 using Moq;
 using Xunit;
 
-namespace Microsoft.AspNet.Mvc.Razor.Test
+namespace Microsoft.AspNet.Mvc.Razor.Compilation
 {
     public class CompilationResultTest
     {
